Refactor install.sh for readability and efficiency

+++ install.sh
@@ -1,69 +1,62 @@
-
 #!/bin/zsh
 
-# Define a function which rename a `target` file to `target.backup` if the file
-# exists and if it's a 'real' file, ie not a symlink
+# Functions
 backup() {
-  target=$1
-  if [ -e "$target" ]; then
-    if [ ! -L "$target" ]; then
-      mv "$target" "$target.backup"
-      echo "-----> Moved your old $target config file to $target.backup"
-    fi
+  local target=$1
+  if [[ -e "$target" && ! -L "$target" ]]; then
+    mv "$target" "$target.backup"
+    echo "-----> Moved your old $target config file to $target.backup"
   fi
 }
 
 symlink() {
-  file=$1
-  link=$2
+  local file=$1
+  local link=$2
   if [ ! -e "$link" ]; then
     echo "-----> Symlinking your new $link"
     ln -s $file $link
   fi
 }
 
-# For all files `$name` in the present folder except `*.sh`, `README.md`, `settings.json`,
-# and `config`, backup the target file located at `~/.$name` and symlink `$name` to `~/.$name`
-# Files removed from this list: irbrc rspec
+install_zsh_plugins() {
+  local plugins_dir="$HOME/.oh-my-zsh/custom/plugins"
+  mkdir -p "$plugins_dir" && cd "$plugins_dir"
+  if [ ! -d "zsh-syntax-highlighting" ]; then
+    echo "-----> Installing zsh plugins..."
+    git clone https://github.com/zsh-users/zsh-autosuggestions
+    git clone https://github.com/zsh-users/zsh-syntax-highlighting
+  fi
+  cd "$1"
+}
+
+# Symlink config files
 for name in aliases gitconfig zprofile zshrc; do
   if [ ! -d "$name" ]; then
-    target="$HOME/.$name"
+    local target="$HOME/.$name"
     backup $target
     symlink $PWD/$name $target
   fi
 done
 
-# Install zsh-syntax-highlighting plugin
-CURRENT_DIR=`pwd`
-ZSH_PLUGINS_DIR="$HOME/.oh-my-zsh/custom/plugins"
-mkdir -p "$ZSH_PLUGINS_DIR" && cd "$ZSH_PLUGINS_DIR"
-if [ ! -d "$ZSH_PLUGINS_DIR/zsh-syntax-highlighting" ]; then
-  echo "-----> Installing zsh plugin 'zsh-syntax-highlighting'..."
-  git clone https://github.com/zsh-users/zsh-autosuggestions
-  git clone https://github.com/zsh-users/zsh-syntax-highlighting
-fi
-cd "$CURRENT_DIR"
+install_zsh_plugins $PWD
 
-# Symlink VS Code settings and keybindings to the present `settings.json` and `keybindings.json` files
-# If it's a macOS
+# Determine OS and set VS Code path accordingly
+code_path=""
 if [[ `uname` =~ "Darwin" ]]; then
-  CODE_PATH=~/Library/Application\ Support/Code/User
-# Else, it's a Linux
+  code_path=~/Library/Application\ Support/Code/User
 else
-  CODE_PATH=~/.config/Code/User
-  # If this folder doesn't exist, it's a WSL
-  if [ ! -e $CODE_PATH ]; then
-    CODE_PATH=~/.vscode-server/data/Machine
-  fi
+  code_path=~/.config/Code/User
+  [[ ! -e $code_path ]] && code_path=~/.vscode-server/data/Machine
 fi
 
+# Symlink VS Code settings
 for name in settings.json keybindings.json; do
-  target="$CODE_PATH/$name"
+  local target="$code_path/$name"
   backup $target
   symlink $PWD/$name $target
 done
 
-# Symlink SSH config file to the present `config` file for macOS and add SSH passphrase to the keychain
+# Handle macOS specifics
 if [[ `uname` =~ "Darwin" ]]; then
   target=~/.ssh/config
   backup $target
@@ -71,20 +64,15 @@ if [[ `uname` =~ "Darwin" ]]; then
   ssh-add -K ~/.ssh/id_ed25519
 fi
 
-# Symlink to zsh theme LENICO
-target="$HOME/.oh-my-zsh/themes/LENICO.zsh-theme"
-symlink $PWD/LENICO.zsh-theme $target
+# Symlink Zsh theme
+symlink $PWD/LENICO.zsh-theme "$HOME/.oh-my-zsh/themes/LENICO.zsh-theme"
 
-# Get the computer name
-computername=$(hostname)
-
-# Check if the computer name starts with "LENICO"
-if [[ $computername == LENICO* ]]; then
-    # If yes, create a symbolic link
+# Check computer name
+if [[ $(hostname) == LENICO* ]]; then
     echo "Computer is LENICO"
 else
     echo "Computer is not LENICO"
 fi
 
-# Refresh the current terminal with the newly installed configuration
+# Activate new configuration
 exec zsh
